requirements.txt
pyodbc==5.2.0
python-dotenv==1.0.1
Requests==2.32.3
schedule==1.2.2
